worker/pyproject.toml
[project]
name = "transcripts-worker"
version = "0.1.0"
description = "Worker service for audio/video transcription"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "dropbox>=12.0.2",
    "pytest>=8.0.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcripts"]